chore: Modified damages to accept extra constructor arguments and override the way to generate the result wrapper
