+++ ra2ce/analysis/losses/multi_link_origin_destination.py
@@ -358,7 +358,7 @@ class MultiLinkOriginDestination(AnalysisBase, AnalysisLossesProtocol):
             self._analysis_input
         ).execute()
         (disruption_impact_df, gdf_ori,) = self.multi_link_origin_destination_impact(
-            gdf, _orod_result_wrapper.analysis_result
+            gdf, _orod_result_wrapper.get_single_analysis()
         )
         try:
             assert self.origins_destinations.region
